\name{demo.knowns}
\docType{data}
\alias{demo.knowns}
\title{Demonstration Knowns Database}

\description{
  A knowns database, for demonstrating the \code{TRAMPR} package.
  This is a subset of a full knowns database, and not intended to
  represent any real data set, and should not be assumed to be
  accurate.

  The data are stored as a \code{\link{TRAMPknowns}} object.  Columns in
  the \code{info} and \code{data} components are described on the
  \code{\link{TRAMPknowns}} page.
}
\usage{data(demo.knowns)}

\section{Licence}{This data set is provided under a Creative Commons
  \dQuote{Attribution-NonCommercial-NoDerivs 2.5} licence.  Please see
  \url{https://creativecommons.org/licenses/by-nc-nd/2.5/} for details.
}

\keyword{datasets}
